replace table creation with alembic
