+++ Dockerfile
@@ -28,3 +28,5 @@ RUN pip install --no-cache-dir --upgrade -r /code/requirements.txt
 RUN apt-get update && apt-get install -y docker.io && rm -rf /var/lib/apt/lists/*
 
 COPY ./src /code/src
+COPY ./migrations /code/migrations
+COPY ./alembic.ini /code/alembic.ini
