+++ docker-compose.yml
@@ -3,7 +3,7 @@ version: '3.8'
 services:
 
   web:
-    build: 
+    build:
       dockerfile: Dockerfile
     ports:
       - 10000:80
@@ -17,7 +17,10 @@ services:
       # non rootless docker
       - /var/run/docker.sock:/var/run/docker.sock
     command:  uvicorn src.api:app --host 0.0.0.0 --port 80 --log-config=src/config/log_config.yaml
+    # run this when there is a database table change
+    #command: "sh -c 'alembic upgrade head && uvicorn src.api:app --host 0.0.0.0 --port 80 --log-config=src/config/log_config.yaml'"
     environment:
       - APSCHEDULE_JOBSTORE_URL=sqlite:////db/scheduler.db
+      - DATABASE_URL=sqlite:////db/database.db
       - TZ=Pacific/Auckland
       - BACKUP_DIR=${BACKUP_DIR}
