+++ src/db.py
@@ -1,20 +1,10 @@
 from typing import Optional
-
+import os
 from sqlmodel import Field, SQLModel, create_engine
+from src.models import *
 
 
-class Backups(SQLModel, table=True):
-    backup_name: Optional[str] = Field(default=None, primary_key=True)
-    backup_path: Optional[str] = Field(default=None)
-    backup_created: str
-    backup_path: str
-    volume_name: str
-    restored: Optional[bool] = Field(default=False)
-    restored_date: Optional[str] = Field(default=None)
-
-
-SQLITE_FILE_NAME = "database.db"
-sqlite_url = f"sqlite:////db/{SQLITE_FILE_NAME}"
+sqlite_url = os.environ.get("DATABASE_URL", f"sqlite://")
 
 engine = create_engine(sqlite_url, echo=True, connect_args={"check_same_thread": False})
